Save java app config and web app config in different tables.

+++ AutoDeployerServer/deploy/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render
 from django.http import HttpResponse
 
-from deploy.models import AppConfig
+from deploy.models import JavaAppConfig
+from deploy.models import WebAppConfig
 
 import json
 import os
@@ -9,21 +10,38 @@ import os
 import Deployer
 
 
+def checkAppType(app_type):
+    result = None
+    if app_type != "java" and app_type != "web":
+        result["code"] = 500
+        result["msg"] = "Unrecognined app type: " + str(app_type)
+        json_str = json.dumps(result)
+
+        result = HttpResponse(json_str)
+    return result
+
+
 def createConfig(request):
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
     config_name = request.GET.get('configName')
     repo_path = request.GET.get('repoPath')
     branch = request.GET.get('branch')
     submodule = request.GET.get('subdir')
-    app_type = request.GET.get('appType')
     conf_path = request.GET.get('conf')
 
-    server_name = request.GET.get('serverName')
-
-    version = request.GET.get('serverName')
+    version = request.GET.get('version')
     if not version:
         version = ""
 
-    configs = AppConfig.objects.filter(config_name=config_name)
+    if app_type == "java":
+        configs = JavaAppConfig.objects.filter(config_name=config_name)
+    elif app_type == "web":
+        configs = WebAppConfig.objects.filter(config_name=config_name)
+
     if len(configs) > 0:
         result = {}
         result["code"] = 500
@@ -32,45 +50,62 @@ def createConfig(request):
 
     print(os.getcwd())
 
-    config = AppConfig(config_name=config_name, repo_path=repo_path, branch=branch, submodule=submodule, app_type=app_type,
-              conf_path=conf_path, server_name=server_name, version=version)
+    invalidte_param_result = {}
+    invalidte_param_result["code"] = 500
+    invalidte_param_result["msg"] = "Invalid parameter."
+
+    if app_type == "java":
+        server_name = request.GET.get('serverName')
+        if not server_name:
+            return HttpResponse(json.dumps(invalidte_param_result))
+
+        config = JavaAppConfig(config_name=config_name, repo_path=repo_path, branch=branch, submodule=submodule,
+                           app_type=app_type, conf_path=conf_path, server_name=server_name, version=version)
+    elif app_type == "web":
+        tomcat_version = request.GET.get('tomcatVersion')
+        if not tomcat_version:
+            return HttpResponse(json.dumps(invalidte_param_result))
+
+        config = WebAppConfig(config_name=config_name, repo_path=repo_path, branch=branch, submodule=submodule,
+                           app_type=app_type,
+                           conf_path=conf_path, tomcat_version=tomcat_version, version=version)
     config.save()
 
-    serialized_config = {}
-    serialized_config["config_name"] = config.config_name
-    serialized_config["repo_path"] = config.repo_path
-    serialized_config["branch"] = config.branch
-    serialized_config["submodule"] = config.submodule
-    serialized_config["app_type"] = config.app_type
-    serialized_config["conf_path"] = config.conf_path
-    serialized_config["server_name"] = config.server_name
-
-    json_str = json.dumps(serialized_config)
+    json_str = config.toJSON()
 
     return HttpResponse(json_str)
 
 
 def list_config(request):
-    all_config_list = AppConfig.objects.order_by('-config_name')
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
+    if app_type == "java":
+        all_config_list = JavaAppConfig.objects.order_by('-config_name')
+    elif app_type == "web":
+        all_config_list = WebAppConfig.objects.order_by('-config_name')
+
     serialized_configs = []
     for config in all_config_list:
-        serialized_config = {}
-        serialized_config["config_name"] = config.config_name
-        serialized_config["repo_path"] = config.repo_path
-        serialized_config["branch"] = config.branch
-        serialized_config["submodule"] = config.submodule
-        serialized_config["app_type"] = config.app_type
-        serialized_config["conf_path"] = config.conf_path
-        serialized_config["server_name"] = config.server_name
-
-        serialized_configs.append(serialized_config)
+        serialized_configs.append(config.toJSON())
     json_str = json.dumps(serialized_configs)
     return HttpResponse(json_str)
 
 
 def delete_config(request):
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
     config_name = request.GET.get('configName')
-    configs = AppConfig.objects.filter(config_name=config_name)
+    if app_type == "java":
+        configs = JavaAppConfig.objects.filter(config_name=config_name)
+    elif app_type == "web":
+        configs = WebAppConfig.objects.filter(config_name=config_name)
+
     print(str(configs))
     for config in configs:
         config.delete()
@@ -83,39 +118,76 @@ def delete_config(request):
 
 
 def deploy(request):
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
     config_name = request.GET.get('configName')
-    config = AppConfig.objects.get(config_name=config_name)
+    if app_type == "java":
+        config = JavaAppConfig.objects.get(config_name=config_name)
+        repoPath = config.repo_path
+        branch = config.branch
+        subdir = config.submodule
+        appType = config.app_type
+        conf = config.conf_path
+        serverName = config.server_name
+        version = config.version
 
-    repoPath = config.repo_path
-    branch = config.branch
-    subdir = config.submodule
-    appType = config.app_type
-    conf = config.conf_path
-    serverName = config.server_name
-    version = config.version
+        result = Deployer.deployAndRunJavaApp(repoPath, branch, subdir, version, appType, conf, serverName)
+    elif app_type == "web":
+        config = WebAppConfig.objects.get(config_name=config_name)
 
-    print(os.getcwd())
+        repoPath = config.repo_path
+        branch = config.branch
+        subdir = config.submodule
+        appType = config.app_type
+        conf = config.conf_path
+        tomcat_version = config.tomcat_version
+        version = config.version
 
-    result = Deployer.deployAndRun(repoPath, branch, subdir, version, appType, conf, serverName)
+        result = Deployer.deployAndRunTomcatApp(repoPath, branch, subdir, version, appType, conf, tomcat_version)
+
     json_str = json.dumps(result)
     return HttpResponse(json_str)
 
 
 def stop(request):
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
     config_name = request.GET.get('configName')
-    config = AppConfig.objects.get(config_name=config_name)
+    if app_type == "java":
+        config = JavaAppConfig.objects.get(config_name=config_name)
+    elif app_type == "web":
+        config = WebAppConfig.objects.get(config_name=config_name)
 
     subdir = config.submodule
     appType = config.app_type
 
-    result = Deployer.stopService(subdir, appType)
+    if app_type == "java":
+        result = Deployer.stopService(subdir, appType, "")
+    elif app_type == "web":
+        tomcatVersion = config.tomcat_version
+        result = Deployer.stopService(subdir, appType, tomcatVersion)
+
     json_str = json.dumps(result)
     return HttpResponse(json_str)
 
 
 def restart(request):
+    app_type = request.GET.get('appType')
+    check_result = checkAppType(app_type)
+    if check_result:
+        return check_result
+
     config_name = request.GET.get('configName')
-    config = AppConfig.objects.get(config_name=config_name)
+    if app_type == "java":
+        config = JavaAppConfig.objects.get(config_name=config_name)
+    elif app_type == "web":
+        config = WebAppConfig.objects.get(config_name=config_name)
 
     subdir = config.submodule
     appType = config.app_type
